services/nalogi/requirements.txt
faust==1.10.4
grpcio==1.29.0
protobuf==3.12.2
